Introduction

Splittings.jl Documentation

Operators splitting package to solve equations of the form

$dU/dt = (T+V)U$,

where $T$ and $V$ are two differential operators by solving successively the simpler equations

The application of an operator splitting method to a concrete problem is done by using Julia macros.

Examples of applications are provided for

The linear pendulum problem.

The Vlasov equation with constant coefficients advection field.

The non linear Vlasov-Poisson equations in cartesian coordinates.

References

E. Hairer, C. Lubich, G. Wanner, Geometrical numerical integration, Springer 2006

This code is derived from Fortran and Python codes written by

- Edwin Chacon Golcher (Institute of Physics of the Czech Academy of Sciences).
- Michel Mehrenberger  (Aix-Marseille Université).
- Eric Sonnendrucker   (Max-Planck-Institut für Plasmaphysik).

Rotation of a gaussian distribution

How to Contribute

Vlasov-Ampere

Bump On Tail

Semi-Lagrangian method

Functions

Index

Splittings.jl Documentation

Examples of applications are provided for

Examples

Functions

Index

Vlasov-Poisson

Semi-Lagrangian method

Examples

Vlasov-Poisson

Vlasov-Ampere

Bump On Tail

Functions

Splittings.compute_rho — Method.

compute_rho( mesh, f)

Compute charge density

ρ(x,t) = ∫ f(x,v,t) dv

return ρ - ρ̄ if neutralized=true

source

Splittings.compute_e — Method.

compute_e( mesh, ρ)

Compute 1d electric field using that -ik * e = ρ

source

Index

Splittings.compute_e

Splittings.compute_rho