Advection functions

Advection functions

Splittings.advection! — Function.

advection!( fᵀ, mesh1, mesh2, E, dt, type, axis )

axis == 1 Advection in x and compute electric field

∂f/∂t − υ∂f/∂x = 0 ∂E/∂t = −J = ∫ fυ dυ

axis == 2 Advection in υ

∂f/∂t − E(x) ∂f/∂υ = 0

source

advection!( mesh, f, v, dt, interp, axis)

Advection of a 2d function f discretized on a 2d mesh along the input axis at velocity v

source

advection!(f, mesh, v, n2, dt, interp)

Advection of a 2d function f along its first dimension with velocity v. Since the fft are computed inplace, the function must be represented by a Array{Complex{Float64},2}.

source

 advection!( f, mesh1, v,  dt)

Semi-lagrangian advection function of 2D distribution function represented by array f. The advection operates along axis (=1 is most efficient) with speed v during dt.

It uses cubic splines interpolation.

source